reorganized example folder
