+++ examples/stream_in_loop.py
